group: relayer wire format + UX fixes (PR-C amend) (#27)
Three fixes folded into PR #26:

1. Relayer wire-format alignment (the big one). Previous wire shape
   was wrong on multiple counts vs. onym-relayer's actual contract
   (`src/handler.rs`):
   * Top-level envelope: now camelCase + adds `contractType` and
     `network` (required by relayer for routing + allowlist check).
     `function` is `create_group`, not the made-up `create_group_v2`.
   * Proof: send the raw 1601-byte PLONK proof, not the
     `Common.parsePlonkProof`-trimmed 1568. Relayer's
     `decode_wire_bytes(_, _, Some(1601))` rejected the trimmed form.
   * Public inputs: send as a JSON array of 4 base64 strings (the
     full 128-byte Tyranny PI bundle split into 4×32B), not as a
     `{commitment, epoch}` object. Relayer's
     `build_public_inputs_from_object → ContractType::Tyranny` arm
     needs all four (commitment + Fr(0) + admin_pubkey_commitment +
     group_id_fr). Also send `admin_pubkey_commitment` as a
     dedicated payload field for the contract's `--admin-pubkey-commitment`
     CLI arg.
   * `SEPGroupType` now `String`-backed ("tyranny", "anarchy", …) to
     match the relayer's lowercase wire labels. PersistedGroup +
     GroupInvitationPayload follow.

2. NetworkPreference + Settings toggle. New `AppNetwork` enum
   (testnet / mainnet) backed by `UserDefaultsNetworkPreference`.
   Settings → Network gets a "Use Mainnet" Toggle bound via
   `@AppStorage(UserDefaultsNetworkPreference.storageKey)`,
   defaulting to off. CreateGroupInteractor pulls the active
   network for both contract-binding lookup AND the wire payload's
   top-level `network` field.

3. UX:
   * Step1 name TextField pre-populates with a random "Adjective
     Noun" placeholder ("Maple Garden" etc) generated at init. On
     first focus the field clears so the user can type; if they
     leave it empty, submit falls back to the placeholder.
     `canAdvanceToStep2` no longer requires a non-empty name.
   * Creating-screen error banner gets a "Cancel" button next to
     "Try again" — closes the whole flow instead of forcing the
     user back to step2 only to dismiss.

Tests: SEPContractClientTests rewritten for the new envelope shape
+ added a mainnet-serializes-as-`public` assertion. GroupProofGenerator
test asserts 1601-byte proof + 4-element PI. CreateGroupInteractor
test adds a network-preference assertion. CreateGroupFlow tests
expanded to cover placeholder generation + first-focus clearing +
effectiveName fallback + tappedCancelFromError.

All affected suites pass on iPhone 17 Pro / iOS 26.

+++ Sources/OnymIOS/Group/SwiftDataGroupStore.swift
@@ -113,7 +113,7 @@ actor SwiftDataGroupStore: GroupStore {
             createdAt: group.createdAt,
             epoch: Int64(bitPattern: group.epoch),
             tierRaw: group.tier.rawValue,
-            groupTypeRaw: Int(group.groupType.rawValue),
+            groupTypeRaw: group.groupType.rawValue,
             isPublishedOnChain: group.isPublishedOnChain,
             encryptedName: try StorageEncryption.encrypt(group.name),
             encryptedGroupSecret: try StorageEncryption.encrypt(group.groupSecret),
@@ -132,7 +132,7 @@ actor SwiftDataGroupStore: GroupStore {
             let members = try? JSONDecoder().decode([GovernanceMember].self, from: membersJSON),
             let salt = try? StorageEncryption.decrypt(row.encryptedSalt),
             let tier = SEPTier(rawValue: row.tierRaw),
-            let groupType = SEPGroupType(rawValue: UInt32(row.groupTypeRaw))
+            let groupType = SEPGroupType(rawValue: row.groupTypeRaw)
         else {
             return nil
         }
